Implemented Battle System
This needs to be cleaned up by a lot, but a battle function IS in and working.
Next steps will probably be a basic save system and then clean up + expand a lot of what's there.
The enemies definitely should be in it's own xml or json file. Wondering if xml would be better for someone hypothetically writing dialogue for a dungeon crawler.